+++ src/main/java/application/service/ConcreteDesignService.java
@@ -1,5 +1,6 @@
 package application.service;
 
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 
 import application.dao.ConcreteDesignDao;
@@ -10,13 +11,18 @@ public class ConcreteDesignService {
 	
 	ConcreteDesignDao dao = DaoFactory.ConcreteDesignlDao();
 	
+	CompresionTestService compresionTestService = new CompresionTestService();
+	
 	public void insertConcreteDesign(ConcreteDesign obj) {
 		dao.insert(obj);
 	}
 	public void updateConcreteDesign(ConcreteDesign obj) {
 		dao.update(obj);
 	}
-	public void deleteConcreteDesignById(Integer id) {
+	public void deleteConcreteDesignById(Integer id) throws SQLIntegrityConstraintViolationException {
+		if (compresionTestService.compresionTestContainsConcreteDesingId(id)) {
+			throw new SQLIntegrityConstraintViolationException("La dosificacíon esta en uso");
+		}
 		dao.deleteById(id);
 	}
 	public ConcreteDesign findConcreteDesignById(Integer id) {
