+++ src/main/java/application/service/ProviderService.java
@@ -1,5 +1,6 @@
 package application.service;
 
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 
 import application.dao.ProviderDao;
@@ -10,6 +11,8 @@ import application.domaim.Provider;
 public class ProviderService {
 
 	private ProviderDao dao = DaoFactory.createProviderDao();
+	
+	private MaterialService materialService = new MaterialService();
 
 	public List<Provider> findAll() {
 		return dao.findAll();
@@ -26,7 +29,10 @@ public class ProviderService {
 		}
 	}
 
-	public void deleteById(Integer id) {
+	public void deleteById(Integer id) throws SQLIntegrityConstraintViolationException {
+		if (materialService.findByProviderId(id).size() > 0) {
+			throw new SQLIntegrityConstraintViolationException("Hay materiales utilizando este proveedor");
+		}
 		dao.deleteById(id);
 	}
